test 1
This is a longer message just to test what that means in a git commit. In this case, I am using only one set of quotes total, hitting enter twice after my short message, and then typing for quite a while to get  multi-line message with the purpose of seeing how it displays in git log or other commands to refer back to a commit. ha;sfjgha;kjdlfhg;aldjfhngkjethgjewthngbipjqetbg[iqjebfkjaefnv;kjaebnfk;jetkgjbnekfjbnaekfjvbna;kejfbnkjaqebfkjaebfvkjbadfkjvbaekjfbvakejfbvkajefbvjkabfv